OH2-475 Stack the dashboard widgets full-width on phone viewports
getBreakpointFromWidth fell back to md below the xs breakpoint, so phone
widths rendered the two-column md layout and squeezed the widgets to
half-width. Map sub-xs widths to xxs (full-width, stacked) instead, and
disable drag/resize/drop on that breakpoint since those interactions are
awkward on touch. Extract the breakpoint helper and its config into a
window-free module and cover the mapping with a unit test.
